+++ playground/zod-pocketbase.config.ts
@@ -4,4 +4,5 @@ import type { Config } from "zod-pocketbase";
 export default {
   ignore: ["users"],
   nameRecordSchema: (name: string) => `z${pascalCase(name)}Record`,
+  output: "./src/lib/pocketbase.ts",
 } satisfies Config;