+++ doc/CHANGELOG.md
@@ -1,5 +1,11 @@
 # doc
 
+## 0.1.0
+
+### Minor Changes
+
+- remove "cache" option for "helpersFrom" and replace it with "fetch" to allow a more generic custom fetch and remove "@11ty/eleventy-fetch" dependency
+
 ## 0.0.2
 
 ### Patch Changes