+++ package/CHANGELOG.md
@@ -1,5 +1,11 @@
 # zod-pocketbase
 
+## 0.5.0
+
+### Minor Changes
+
+- remove "cache" option for "helpersFrom" and replace it with "fetch" to allow a more generic custom fetch and remove "@11ty/eleventy-fetch" dependency
+
 ## 0.4.1
 
 ### Patch Changes